+++ lib/phoenix_kit_warehouse/web/goods_issue_form_live.ex
@@ -1106,15 +1106,13 @@ defmodule PhoenixKitWarehouse.Web.GoodsIssueFormLive do
                     phx-submit="set_issued_qty"
                   >
                     <input type="hidden" name="index" value={index} />
-                    <input
-                      type="number"
+                    <.decimal_input
                       id={"gi-iss-#{index}"}
                       name="issued_quantity"
-                      min="0"
-                      step="any"
                       value={fmt_qty(line["issued_quantity"])}
                       placeholder="0"
-                      class="input input-sm w-24 text-right tabular-nums"
+                      class="input-sm text-right tabular-nums"
+                      wrapper_class="inline-block w-24"
                       phx-debounce="blur"
                       phx-hook="InvEnterBlur"
                     />